requirements.txt
aiohttp==3.7.3
aiohttp-jinja2==1.4.2
async-timeout==3.0.1
attrs==20.3.0
bidict==0.21.2
certifi==2020.12.5
chardet==3.0.4
dnspython==1.16.0
eventlet==0.30.1
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
numpy==1.20.1
python-engineio==4.0.0
python-socketio==5.0.4
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
websockets==8.1
yarl==1.6.3

python-dotenv~=0.15.0
